+++ robotcode/language_server/robotframework/protocol.py
@@ -29,6 +29,7 @@ from .parts.folding_range import RobotFoldingRangeProtocolPart
 from .parts.formatting import RobotFormattingProtocolPart
 from .parts.goto import RobotGotoProtocolPart
 from .parts.hover import RobotHoverProtocolPart
+from .parts.inlay_hint import RobotInlayHintProtocolPart
 from .parts.inline_value import RobotInlineValueProtocolPart
 from .parts.references import RobotReferencesProtocolPart
 from .parts.rename import RobotRenameProtocolPart
@@ -94,6 +95,7 @@ class RobotLanguageServerProtocol(LanguageServerProtocol):
     robot_selection_range = ProtocolPartDescriptor(RobotSelectionRangeProtocolPart)
     robot_rename = ProtocolPartDescriptor(RobotRenameProtocolPart)
     robot_inline_value = ProtocolPartDescriptor(RobotInlineValueProtocolPart)
+    robot_inlay_hint = ProtocolPartDescriptor(RobotInlayHintProtocolPart)
     robot_code_action = ProtocolPartDescriptor(RobotCodeActionProtocolPart)
     robot_workspace = ProtocolPartDescriptor(RobotWorkspaceProtocolPart)